+++ authlete/dto/token_create_request.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2019 Authlete, Inc.
+# Copyright (C) 2019-2020 Authlete, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -33,7 +33,9 @@ class TokenCreateRequest(Jsonable):
             'clientIdAliasUsed':     bool,
             'accessToken':           str,
             'refreshToken':          str,
-            'accessTokenPersistent': bool
+            'accessTokenPersistent': bool,
+            'certificateThumbprint': str,
+            'dpopKeyThumbprint':     str,
         }
 
         super().__init__(nameAndValues, nameAndTypes)